del old util test
