+++ packages/repo/tests/car.test.ts
@@ -20,4 +20,20 @@ describe('car', () => {
 
     await readCarStream(car)
   })
+
+  it('writeCar propagates errors', async () => {
+    const iterate = async () => {
+      async function* blockIterator() {
+        await wait(1)
+        const block = await dataToCborBlock({ test: 1 })
+        yield block
+        throw new Error('Oops!')
+      }
+      const iter = writeCarStream(null, blockIterator())
+      for await (const _bytes of iter) {
+        // no-op
+      }
+    }
+    await expect(iterate).rejects.toThrow('Oops!')
+  })
 })
